+++ src/utils/l10n/flattenMessages.ts
